services/event-tracker/requirements.txt
auth0-python==3.21.0
alembic==1.7.1
appdirs==1.4.4
atomicwrites==1.4.0
attrs==21.2.0
aws-xray-sdk==2.5.0
black==20.8b1
boto3==1.9.142
botocore==1.12.142
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
click==8.0.1
cognitojwt==1.4.1
cryptography==3.4.8
docutils==0.17.1
ecdsa==0.17.0
factory-boy==3.2.0
Faker==8.14.0
Flask==1.0.2
Flask-Cognito==1.13
Flask-Cors==3.0.8
Flask-GraphQL==2.0.1
Flask-Login==0.4.1
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.4.0
future==0.18.2
graphene==2.1.8
graphene-sqlalchemy==2.2.2
idna==2.8
importlib-metadata==4.8.1
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
jsonpickle==2.0.0
Mako==1.1.5
MarkupSafe==2.0.1
more-itertools==8.9.0
mypy-extensions==0.4.3
pathspec==0.9.0
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.10.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.2.0
pytest==3.8.1
pytest-flask==0.12.0
pytest-mock==2.0.0
python-dateutil==2.8.2
python-jose==3.3.0
regex==2021.8.28
requests==2.21.0
rsa==4.7.2
s3transfer==0.2.1
six==1.16.0
SQLAlchemy==1.3.23
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.2
Werkzeug=2.0.3
urllib3==1.24.3
uuid==1.30
wrapt==1.12.1
zipp==3.5.0
